backend/requirements.txt
attrs==17.3.0
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
falcon==1.3.0
idna==2.6
msgpack-python==0.4.8
munch==2.2.0
pluggy==0.6.0
py==1.5.2
Pygments==2.2.0
python-mimeparse==1.6.0
requests==2.18.4
six==1.11.0
urllib3==1.22
waitress==1.1.0
